+++ swarm-controller-service/src/test/java/io/pockethive/swarmcontroller/SwarmControllerTestProperties.java
@@ -44,7 +44,7 @@ final class SwarmControllerTestProperties {
                         METRICS_SHUTDOWN_OPERATION,
                         METRICS_JOB,
                         new SwarmControllerProperties.GroupingKey(METRICS_GROUPING_INSTANCE))),
-                new SwarmControllerProperties.Docker(null, "/var/run/docker.sock"),
+                new SwarmControllerProperties.Docker(null, "/var/run/docker.sock", null),
                 new SwarmControllerProperties.Features(bufferGuardEnabled)));
     }
 }